feat: implement operation chaining & proper equation staging
This commit allows users to correctly perform operations after a successful equation.
To be more specific, in the event that after the user has clicked the 'equals' button,
they can use the result of the equation as the first value for another equation.
This function can be replicated upon using an "operator" symbol as the equation result
has been previously displayed. On another note, this commit also changes the way equation staging works.
Specifically, left hand operation strings  will now be displayed on the "equation-stage" container while
results are highlighted over the "entry-field" container. This provides a clearer and familiar usage
over the calculator application.